- reverted test for executors to be more verbose. - test_main.cpp added to be used by multiple test applications.
