+++ test/CMakeLists.txt
@@ -1,11 +1,12 @@
 # SPDX-License-Identifier: Unlicense
 # SPDX-FileCopyrightText: 2023 NeoFOAM authors
 
-add_executable(test_field "test_Field.cpp")
+add_executable(test_field "test_Field.cpp" "test_main.cpp")
 target_link_libraries(test_field PRIVATE Catch2::Catch2 NeoFOAM Kokkos::kokkos)
 
-add_executable(test_executor "test_Executor.cpp")
-target_link_libraries(test_executor PRIVATE Catch2::Catch2 NeoFOAM Kokkos::kokkos)
+add_executable(test_executor "test_Executor.cpp" "test_main.cpp")
+target_link_libraries(test_executor PRIVATE Catch2::Catch2 NeoFOAM
+                                            Kokkos::kokkos)
 
 list(APPEND CMAKE_MODULE_PATH ${catch2_SOURCE_DIR}/extras)
